+++ internal/xiter/xiter_experiment_go1.22.go
@@ -1,0 +1,9 @@
+//go:build go1.22 && goexperiment.rangefunc
+
+package xiter
+
+import (
+	"iter"
+)
+
+type Seq2[K, V any] iter.Seq2[K, V]