+++ query/result.go
@@ -6,6 +6,7 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/closer"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/scanner"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/types"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xiter"
 )
 
 type (
@@ -14,11 +15,17 @@ type (
 
 		NextResultSet(ctx context.Context) (ResultSet, error)
 		Err() error
+
+		// Range is experimental API for range iterators available with Go version 1.22+ and flag `GOEXPERIMENT=rangefunc`.
+		Range(ctx context.Context) xiter.Seq2[ResultSet, error]
 	}
 	ResultSet interface {
 		Columns() []string
 		ColumnTypes() []Type
 		NextRow(ctx context.Context) (Row, error)
+
+		// Range is experimental API for range iterators available with Go version 1.22+ and flag `GOEXPERIMENT=rangefunc`.
+		Range(ctx context.Context) xiter.Seq2[Row, error]
 	}
 	Row interface {
 		Scan(dst ...interface{}) error