+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added experimental range functions to the `query.Result` and `query.ResultSet` types, available as for-range loops starting with Go version 1.22. These features can be enabled by setting the environment variable `GOEXPERIMENT=rangefunc`.
 * Added public types for `tx.Option`, `options.DoOption` and `options.DoTxOption`
 
 ## v3.73.1